Extract crashTrend helper
Move the crash-trend calculation into its own file so `ReleaseHealth+Summaries` stays focused on summary aggregation.
